move interactive function out of index
